# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nguyen Trong Doanh",
        "title": "API Echo_Swagger",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1234",
    "basePath": "/v2",
    "paths": {
        "/users": {
            "get": {
                "description": "Trả về danh sách tất cả người dùng từ cơ sở dữ liệu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Lấy danh sách người dùng",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.User"
                            }
                        }
                    }
                }
            }
        },
        "/users/create": {
            "post": {
                "description": "Tạo một người dùng mới với thông tin được cung cấp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Tạo người dùng mới",
                "parameters": [
                    {
                        "description": "Thông tin người dùng mới",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Người dùng đã được tạo thành công",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                }
            }
        },
        "/users/delete/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Xóa người dùng",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của người dùng cần xóa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Người dùng đã bị xóa",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/update/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Cập nhật thông tin người dùng",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của người dùng cần cập nhật",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin người dùng cần cập nhật",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.User": {
            "description": "Nguyen Trong Doanh",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}